image clone controller: initial version
This commit adds an initial version of the image clone controller.

The controller watches Deployments and DaemonSets and copies the images
to the destination registry and updates the image field to the
destination repository.

cli - contains the code related to cli
pkg - contains the code related to registry, manager and controllers
deploy - contains the manifests required to run image clone controller.

Signed-off-by: Imran Pochi <official.immi@gmail.com>